Implement BLoC to Film Detail Page
