+++ movie_shape/lib/views/film_detail_page/state/bloc/film_detail_page_event.dart
@@ -1,0 +1,11 @@
+part of 'film_detail_page_bloc.dart';
+
+abstract class FilmDetailPageEvent {
+  const FilmDetailPageEvent();
+}
+
+class LoadFilm extends FilmDetailPageEvent {
+  final String filmId;
+
+  const LoadFilm(this.filmId);
+}
