+++ movie_shape/lib/views/nav_rail/state/nav_rail_bloc.dart
@@ -4,7 +4,7 @@ import 'package:meta/meta.dart';
 import 'package:movie_shape/models/film.dart';
 import 'package:movie_shape/views/homepage/home_page.dart';
 import 'package:movie_shape/views/favourites/favourites_page.dart';
-import 'package:movie_shape/views/pages/film_detail_page.dart';
+import 'package:movie_shape/views/film_detail_page/film_detail_view.dart';
 import 'package:movie_shape/views/watchlist/watchlist_page.dart';
 
 import '../../../repository/film_repo_implemented.dart';
@@ -27,7 +27,6 @@ class NavRailBloc extends Bloc<NavRailEvent, NavRailState> {
           emit(DisplaySelectedPage(event.selectedIndex, FavouritesPage()));
         case 2:
           emit(DisplaySelectedPage(event.selectedIndex, WatchlistPage()));
-        case 3:
         default:
           emit(NavRailError('no page found'));
       }
