+++ movie_shape/lib/views/watchlist/watchlist_view.dart
@@ -4,7 +4,8 @@ import 'package:movie_shape/models/film.dart';
 import 'package:movie_shape/repository/film_repo_implemented.dart';
 import 'package:movie_shape/reusable/list/list_card.dart';
 import 'package:movie_shape/reusable/page_title_card.dart';
-import 'package:movie_shape/views/pages/film_detail_page.dart';
+import 'package:movie_shape/views/film_detail_page/film_detail_view.dart';
+import 'package:movie_shape/views/nav_rail/state/nav_rail_bloc.dart';
 import 'package:movie_shape/views/watchlist/state/watchlist_bloc/watchlist_bloc.dart';
 import 'package:movie_shape/views/widgets/searchbar.dart';
 
@@ -13,15 +14,15 @@ class WatchlistView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    Future<void> _goToDetailPage(String id) async {
-      Film film = await FilmRepoImplemented().getFilmById(id: id);
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (context) => FilmDetailPage(film: film),
-        ),
-      );
-    }
+    // Future<void> _goToDetailPage(String id) async {
+    //   Film film = await FilmRepoImplemented().getFilmById(id: id);
+    //   Navigator.push(
+    //     context,
+    //     MaterialPageRoute(
+    //       builder: (context) => FilmDetailView(film: film),
+    //     ),
+    //   );
+    // }
 
     return Column(
       children: [
@@ -66,8 +67,8 @@ class WatchlistView extends StatelessWidget {
                     final film = state.watchlist[index];
                     return ListCard(
                       film: film,
-                      onTap: (val) {
-                        _goToDetailPage(val);
+                      onTap: (val) async {
+                        context.read<NavRailBloc>().add(FilmSelected(val));
                       },
                     );
                   },
